+++ src/main/java/project/view/servlet/RegisterFormServlet.java
@@ -1,4 +1,4 @@
-package project.view;
+package project.view.servlet;
 
 
 import javax.servlet.http.HttpServlet;
